easy coding -
recap & re-organized about 泛型约束

#1 赋值操作 - 只能接受符合泛型约束的变量
#2 add()操作 - extends不支持add(适合仅get的场合), super支持add T与T的子类型(适合于put场合)
