--- src/main/java/com/henry/tryout/easy_coding/data_structure_and_collection/collection_and_generic_05/Animal.java
@@ -1,5 +1,0 @@
-package com.henry.tryout.easy_coding.data_structure_and_collection.collection_and_generic_05;
-
-public class Animal {
-
-}
