+++ src/main/java/com/henry/tryout/easy_coding/data_structure_and_collection/collection_and_generic_05/multiple_type_under_constraints_03/Cat.java
@@ -1,0 +1,5 @@
+package com.henry.tryout.easy_coding.data_structure_and_collection.collection_and_generic_05.multiple_type_under_constraints_03;
+
+public class Cat extends Animal {
+
+}
